requirements.txt
keyring
psycopg2-binary
redis
pymongo
cassandra-driver
flask
requests
pytest